Feat: alteração feita no allies sheet holder.hpp

+++ src/model/character/include/AlliesSheetHolder.hpp
@@ -1,12 +1,16 @@
+
 #ifndef ALLIES_SHEET_HOLDER_HPP
 #define ALLIES_SHEET_HOLDER_HPP
 
+#include <string>
 #include <vector>
-#include "TechnicalSheet.hpp"
+#include <memory>
+#include <stdexcept>
+#include "./include/TechnicalSheet.hpp"
 
 class AlliesSheetHolder {
 private:
-    const int maxAllies = 3;
+    const size_t maxAllies = 3;
     std::vector<TechnicalSheet> partyMembers;
 
 public:
@@ -17,12 +21,12 @@ public:
     std::vector<TechnicalSheet>& getAllies();
 
     // Deletar ficha de aliado
-    void deleteAllies(int indice);
+    void deleteAllies(size_t indice);
 
     // Adicionar aliado
     void addAllies(const TechnicalSheet& newAllie);
 
-    // Retorna o número de aliados
+    // Verifica quantidade de aliados
     size_t getNumAllies() const;
 };
